pyproject.toml
[build-system]
requires = [
    "setuptools>=42",
    "wheel"
]
build-backend = "setuptools.build_meta"

[project]
requires-python = ">=3.10"
dependencies = [
    "aioredis >= 2.0.0, < 2.1.0",
    "jsonpickle >= 3.0.0, < 3.1.0",
    "tomlkit >= 0.11.0, < 0.12.0",
    "typer >= 0.4.0, < 0.5.0",
]
name = "aiotaskq"
version = "0.0.14"
readme = "README.md"
description = "A simple asynchronous task queue"
authors = [
    {name = "Imran Ariffin", email = "ariffin.imran@gmail.com"},
]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
license = { file = "LICENSE" }

[project.optional-dependencies]
dev = [
    "black >= 22.2.0, < 23.0.0",
    "celery >= 5.4.0, < 5.5.0",
    "coverage >= 6.4.0, < 6.5.0",
    "django-stubs >= 5.1.0, < 5.2.0",
    "mypy >= 0.931, < 1.0",
    "mypy-extensions >= 0.4.0, < 0.5.0",
    "pytest-asyncio >= 0.19.0, < 0.20.0",
    "pylint >= 2.14.0, < 2.15.0",
    "psycopg2 >= 2.9.0, < 2.10.0",
    "pytest >= 7.1.0, < 7.2.0",
    "redis >= 5.2.0, < 5.3.0",
    "typing_extensions >= 4.11.0, < 4.12.0",
]

[project.urls]
"Bug Tracker" = "https://github.com/imranariffin/aiotaskq/issues"

[project.scripts]
aiotaskq = "aiotaskq.__main__:cli"

[tool.setuptools.packages.find]
where = ["src"]
include = ["aiotaskq*"] 

[tool.black]
line-length = 100

[tool.pytest.ini_options]
markers = [
    "lowlevel: Marks tests as low-level",
]
